+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\User\\Desktop\\todos-trello\\src\\index.js":"1","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\App.js":"2","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\index.js":"3","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Content.js":"4","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Header.js":"5","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\actions.js":"6","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Footer.js":"7","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\reducer.js":"8","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Checking.js":"9","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Button\\index.jsx":"10"},{"size":407,"mtime":1611504347457,"results":"11","hashOfConfig":"12"},{"size":708,"mtime":1611504347442,"results":"13","hashOfConfig":"12"},{"size":337,"mtime":1611504347457,"results":"14","hashOfConfig":"12"},{"size":953,"mtime":1611506745316,"results":"15","hashOfConfig":"12"},{"size":173,"mtime":1611504347442,"results":"16","hashOfConfig":"12"},{"size":1414,"mtime":1611506679887,"results":"17","hashOfConfig":"12"},{"size":316,"mtime":1611504347442,"results":"18","hashOfConfig":"12"},{"size":1607,"mtime":1611506623918,"results":"19","hashOfConfig":"12"},{"size":763,"mtime":1611506824787,"results":"20","hashOfConfig":"12"},{"size":684,"mtime":1611506584223,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1dso9rw",{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\User\\Desktop\\todos-trello\\src\\index.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\App.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\index.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Content.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Header.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\actions.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Footer.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\reducer.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Checking.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Button\\index.jsx",[]]
+[{"C:\\Users\\User\\Desktop\\todos-trello\\src\\index.js":"1","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\App.js":"2","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\index.js":"3","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Todos.js":"4","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Header.js":"5","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\actions.js":"6","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Footer.js":"7","C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\todos.js":"8","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Checking.js":"9","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Button\\index.jsx":"10","C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Todo.js":"11"},{"size":407,"mtime":1611504347457,"results":"12","hashOfConfig":"13"},{"size":262,"mtime":1611575334254,"results":"14","hashOfConfig":"13"},{"size":415,"mtime":1611574566755,"results":"15","hashOfConfig":"13"},{"size":1179,"mtime":1611583104630,"results":"16","hashOfConfig":"13"},{"size":170,"mtime":1611573826296,"results":"17","hashOfConfig":"13"},{"size":1313,"mtime":1611507128017,"results":"18","hashOfConfig":"13"},{"size":316,"mtime":1611504347442,"results":"19","hashOfConfig":"13"},{"size":1517,"mtime":1611574566708,"results":"20","hashOfConfig":"13"},{"size":653,"mtime":1611582887683,"results":"21","hashOfConfig":"13"},{"size":596,"mtime":1611579829805,"results":"22","hashOfConfig":"13"},{"size":182,"mtime":1611575334191,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"1dso9rw",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"C:\\Users\\User\\Desktop\\todos-trello\\src\\index.js",[],["47","48"],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\App.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\index.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Todos.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Header.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\actions.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Footer.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\redux\\todos.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Checking.js",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Button\\index.jsx",[],"C:\\Users\\User\\Desktop\\todos-trello\\src\\components\\Todo.js",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},"no-native-reassign",["53"],"no-negated-in-lhs",["54"],"no-global-assign","no-unsafe-negation"]
