+++ debug.log
@@ -1,1 +1,4 @@
 [0124/140632.545:ERROR:directory_reader_win.cc(43)] FindFirstFile: Системе не удается найти указанный путь. (0x3)
+[0124/185718.723:ERROR:directory_reader_win.cc(43)] FindFirstFile: Системе не удается найти указанный путь. (0x3)
+[0124/231055.744:ERROR:filesystem_win.cc(129)] GetFileAttributes C:\Users\User\AppData\Local\Google\Chrome\User Data\Crashpad\attachments\7d2d2196-ee07-4fcf-a5bc-25b973cb7bbe: Системе не удается найти указанный путь. (0x3)
+[0125/135907.771:ERROR:directory_reader_win.cc(43)] FindFirstFile: Системе не удается найти указанный путь. (0x3)
