+++ src/components/Header.js
@@ -1,11 +1,7 @@
 import React from 'react';
 
 function Header(props) {
-  return (
-    <div className="header">
-      СПИСОК ДЕЛ:
-    </div>
-  );
+  return <div className="header">СПИСОК ДЕЛ:</div>;
 }
 
 export default Header;
