+++ src/components/Todo.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Todo(props) {
+  return <div className="title">{props.todo.title}</div>;
+}
+
+export default Todo;
