+++ src/redux/index.js
@@ -1,13 +1,17 @@
-import { applyMiddleware, createStore } from 'redux';
+import { applyMiddleware, combineReducers, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { createLogger } from 'redux-logger';
-import reducer from './reducer'
+import todosReducer from './todos';
 
 const logger = createLogger({
   diff: true,
   collapsed: true,
 });
 
-const store = createStore(reducer, applyMiddleware(thunk, logger));
+const rootReducer = combineReducers({
+  todos: todosReducer,
+});
+
+const store = createStore(rootReducer, applyMiddleware(thunk, logger));
 
 export default store;
